backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
APScheduler==3.10.4
python-dotenv==1.0.1
twilio==9.3.3
python-dateutil==2.9.0.post0
pytz==2024.1
